Implement Unit of Work pattern
Create IUnitOfWork interface and an implementation.
Use UnitOfWork to have access to all repositories and so that all Controllers will have to only know about the UnitOfWork, not the repositories themselves.
This way, we can also ensure that all repositories use the same database context.

Commit and Dispose methods are not used because I made a decision to do the savings (for create, update and delete) in the GenericRepository.
The other way would be to use Save and Dispose methods in the Controller, calling the IUnitOfWork's Commit/CommitAsync method when creating, updating or deleting and use IUnitOfWork'S Dispose when we dispose the Controller itself (override method of MVC Controller type).

More info: https://learn.microsoft.com/en-us/aspnet/mvc/overview/older-versions/getting-started-with-ef-5-using-mvc-4/implementing-the-repository-and-unit-of-work-patterns-in-an-asp-net-mvc-application,
https://stackoverflow.com/questions/55239380/why-derive-from-controllerbase-vs-controller-for-asp-net-core-web-api,
https://learn.microsoft.com/en-us/aspnet/core/web-api/?view=aspnetcore-7.0